# Conflicts:
#	FamTec/FamTec/.vs/FamTec/DesignTimeBuild/.dtbcache.v2
#	FamTec/FamTec/.vs/FamTec/v17/.futdcache.v2
#	FamTec/FamTec/.vs/FamTec/v17/.suo
#	FamTec/FamTec/.vs/FamTec/v17/DocumentLayout.json
#	FamTec/FamTec/.vs/FamTec/v17/fileList.bin
#	FamTec/FamTec/.vs/ProjectEvaluation/famtec.metadata.v7.bin
#	FamTec/FamTec/.vs/ProjectEvaluation/famtec.projects.v7.bin
#	FamTec/FamTec/FamTec/Client/Middleware/ApiManager.cs
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/FamTec.Client.dll
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/FamTec.Client.pdb
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/FamTec.Shared.dll
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/FamTec.Shared.pdb
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Client.dll
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Client.dll.gz
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Client.pdb
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Client.pdb.gz
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Shared.dll
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Shared.dll.gz
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Shared.pdb
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/FamTec.Shared.pdb.gz
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/blazor.boot.json
#	FamTec/FamTec/FamTec/Client/bin/Debug/net7.0/wwwroot/_framework/blazor.boot.json.gz
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.AssemblyInfo.cs
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.AssemblyInfoInputs.cache
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.GeneratedMSBuildEditorConfig.editorconfig
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.csproj.AssemblyReference.cache
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.csproj.FileListAbsolute.txt
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.dll
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.pdb
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/FamTec.Client.sourcelink.json
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/blazor.boot.json
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/compressed/e787y28gy0.gz
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/compressed/eawb6f54y0.gz
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/compressed/jh7rs5n8nq.gz
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/compressed/sl4d152w7g.gz
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/compressed/v5xmy37558.gz
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/ref/FamTec.Client.dll
#	FamTec/FamTec/FamTec/Client/obj/Debug/net7.0/refint/FamTec.Client.dll
#	FamTec/FamTec/FamTec/Client/wwwroot/index.html
#	FamTec/FamTec/FamTec/Shared/bin/Debug/net7.0/FamTec.Shared.dll
#	FamTec/FamTec/FamTec/Shared/bin/Debug/net7.0/FamTec.Shared.pdb
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/FamTec.Shared.AssemblyInfo.cs
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/FamTec.Shared.AssemblyInfoInputs.cache
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/FamTec.Shared.dll
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/FamTec.Shared.pdb
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/FamTec.Shared.sourcelink.json
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/ref/FamTec.Shared.dll
#	FamTec/FamTec/FamTec/Shared/obj/Debug/net7.0/refint/FamTec.Shared.dll

+++ FamTec/FamTec/FamTec/Shared/Client/DTO/Normal/Material/Detail/DetailMaterialListDTO.cs
@@ -48,7 +48,7 @@ namespace FamTec.Shared.Client.DTO.Normal.Material.Detail
         /// <summary>
         /// 총 재고 수량
         /// </summary>
-        public long TotalStockNum { get; set; }
+        public int TotalStockNum { get; set; }
 
         /// <summary>
         /// 이월재고 수량
